increased multipart body length

+++ src/Tetrifact.Web/Startup.cs
@@ -40,7 +40,7 @@ namespace Tetrifact.Web
             {
                 // SECURITY WARNING : the limit on attachment part size is removed to support large
                 // builds. 
-                options.MultipartBodyLengthLimit = int.MaxValue;
+                options.MultipartBodyLengthLimit = long.MaxValue;
             });
 
             // do all ninject stuff here
